Avoid modifying sim state with additional fields
Apparently adding the command field in the sim_state breaks integration of
the sim with the UI. In the train tab, state variables have weird numbers
(for instance, double, which should be a Boolean, has values ranging from 0
to ~30).
It is not clear whether this has implications on the model training or if
it's only a UI glitch.